devtools/client/webconsole/new-console-output/test/components/test_console-api-call.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for ConsoleApiCall component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for ConsoleApiCall component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { BrowserLoader } = Components.utils.import(
    "resource://devtools/client/shared/browser-loader.js", {});
  const rootUrl = "resource://devtools/client/shared/components/";
  const require = BrowserLoader({
    baseURI: rootUrl,
    window: this}).require;

  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/actions/messages");

  const { ConsoleApiCall } = browserRequire("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");

  const packet = yield getPacket("console.log('foobar', 'test')", "consoleAPICall");
  const message = prepareMessage(packet);
  const rendered = renderComponent(ConsoleApiCall, {message});

  const queryPath = "span span.message-flex-body span.message-body.devtools-monospace span.console-string";
  is(1, rendered.querySelectorAll(queryPath).length, "ConsoleApiCall component produces expected DOM")

  SimpleTest.finish()
});
</script>
</body>
</html>
